// src/api-serverless/src/generated/models/WaveScope.ts
/**
 * Seize API
 * This is the API interface description for the most commonly used operations in Seize API.  Some modifying endpoints require an authentication token.   We are in the process of documenting all Seize APIs.   If there is an API that you need, please ping us in Discord and we will aim to prioritize its documentation.
 *
 * OpenAPI spec version: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { Curation } from '../models/Curation';
import { WaveScopeType } from '../models/WaveScopeType';

/**
 * A scope of audience
 */
export class WaveScope {
  'type': WaveScopeType;
  'group': Curation | null;

  static readonly discriminator: string | undefined = undefined;

  static readonly attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
    format: string;
  }> = [
    {
      name: 'type',
      baseName: 'type',
      type: 'WaveScopeType',
      format: ''
    },
    {
      name: 'curation',
      baseName: 'curation',
      type: 'Curation',
      format: ''
    }
  ];

  static getAttributeTypeMap() {
    return WaveScope.attributeTypeMap;
  }

  public constructor() {}
}